+++ js/app/components.js
@@ -22,6 +22,7 @@ const BUILTIN_COMPONENTS = [
   require('../panel/navigator/actions'),
   require('../panel/monitor/stats'),
   require('../panel/monitor/table'),
+  require('../panel/monitor/gauge'),
   require('../panel/navigator/history'),
   require('../panel/navigator/groups'),
   require('../panel/fabric/ports-pane'),   // component-ports (dataflow fabric, P1.5)